% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate}
\usage{
aggregate(
  query,
  clusterid,
  assay = "counts",
  size.factor = 1,
  pool.factor = 1,
  nPool = 15
)
}
\arguments{
\item{query}{Required. Query sce.}

\item{clusterid}{Required. The query Coldata attribute name. Cells will be aggregated within the clusters of 'clusterid'.}

\item{assay}{Default: 'counts'. On which assay of the query we perform aggregation.}

\item{size.factor}{Default: 1. Determine how many pseudo-bulk samples to be generated from each cluster. The number equals to "size.factor times cluster size".}

\item{pool.factor}{Default: 1.  Determine how many cells to be pooled for aggregation in each cluster. The number equals to "pool.factor times cluster size"}

\item{nPool}{Default: NULL. Pooling size of each pseudo-bulk sample. Will suppress pool.factor if specified.}
}
\value{
A confusion matrix if trueAnno has been provided.
}
\description{
Aggregate single cells to create pseudo-bulk samples.
}
