% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSce.R
\name{createSce}
\alias{createSce}
\title{Create SingleCellExperiment Object}
\usage{
createSce(counts = NULL, data = NULL, colData = NULL, as.sparse = TRUE)
}
\arguments{
\item{data}{Optional. A gene by cell expression matrix (can be normalized count).}

\item{infile}{Optional. A gene by cell count matrix.}
}
\value{
SingleCellExperiment Object.
}
\description{
This function converts single cell expression data matrix and metadata to a
SingleCellExperiment object (sce).
}
