% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffusionReconstruct.R
\name{DiffusionReconstruct}
\alias{DiffusionReconstruct}
\title{Diffusion Reconstruction}
\usage{
DiffusionReconstruct(
  reference,
  query,
  Npc,
  a = 2,
  returnSce = TRUE,
  returnDiffuseObj = FALSE,
  colReference.by = NULL,
  colQuery.by = NULL,
  referenceColors = NULL,
  queryColors = NULL,
  referenceText = NULL,
  queryText = NULL,
  ...
)
}
\arguments{
\item{reference}{Required. Reference sce.}

\item{query}{Required. Query sce.}

\item{Npc}{Default: all. How many principle components to use for diffusion map.}

\item{a}{Default: 2. The diffusion bandwidth will be set to 'a' times the maximum distance between reference sample paires.}

\item{returnSce}{Default: TRUE. Whether to return sce objects.}

\item{returnDiffuseObj}{Default: FALSE. Whether to return the diffusion map model.}

\item{colReference.by}{Optional. Color reference samples by which reference metadata attribute.}

\item{colQuery.by}{Optional. Color quey samples by which query metadata attribute.}

\item{referenceColors}{Optional. Color Scheme of colReference.by Should be a named vector using color codes as values and labels of 'colReference.by' as names.}

\item{queryColors}{Optional. Color Scheme of colQuery.by Should be a named vector using color codes as values and labels of 'colQuery.by' as names.}

\item{referenceText}{Optional. Additional annotation of reference samples}

\item{queryText}{Optional. Additional annotation of query cells.}

\item{nPool}{Optional. Pooling size of each pseudo-bulk sample. Will suppress pool.factor if specified.}
}
\value{
A plotly object.
}
\description{
Diffusion reconstruction of the PCA projection landscape.
}
