% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureWeighting.R
\name{featureWeighting}
\alias{featureWeighting}
\title{Feature Weighting}
\usage{
featureWeighting(
  reference,
  clusterid,
  cut = T,
  n.cut = 2000,
  assay2rank = "data",
  rowDataPrefix = "HD_"
)
}
\arguments{
\item{reference}{Required. Reference sce}

\item{clusterid}{Required. Variable name in the metadata that stores sample labels.}

\item{cut}{Default: TRUE Whether to filter the reference data to the gene sets with the 'n.cut' highest HD scores.}

\item{n.cut}{Default: 2000. How many genes to keep in the reference data.}

\item{assay2rank}{Default: 'data'. On the data of which assay we perform rank transformation and then HD calculation?}

\item{rowDataPrefix}{Default: 'HD_'. Prefix added to the column names of HD scores, which will be stored in the RowData of the reference sce.}
}
\value{
SingleCellExperiment Object containing HD scores in the metadata and the RowData.
}
\description{
Calculate Hellinger Distance (HD) between K-Mean discretized gene expression and reference
sample label. Genes with high HD are discriminant against sample labels. We can control whether
to filter the reference data by HD scores using the 'cut' and the 'n.cut' option.
}
