% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterData.R
\name{filterData}
\alias{filterData}
\title{Filter data}
\usage{
filterData(
  reference,
  query,
  min.cell = 10,
  max.sparsity = 0.98,
  referenceAssay = "rank",
  queryAssay = "data"
)
}
\arguments{
\item{reference}{Required. Reference sce.}

\item{query}{Required. Query sce.}

\item{min.cell}{Default: 10. Query genes expressed in less then 'min.cell' cells are filtered.}

\item{max.sparsity}{Default: 0.98. Cells with proportion of zeros larger than 'max.sparsity' are filtered.}

\item{referenceAssay}{Default: 'rank'. On which reference assay we perform rank transformation.}

\item{queryAssay}{Default: 'data'. On which query assay we perform quality controls.}
}
\value{
reference and query sce objects.
}
\description{
Quality control the query data and filter the reference and query to their
common gene set.
}
