% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HellingerDist.R
\name{HellingerDist}
\alias{HellingerDist}
\title{Hillinger Distance}
\usage{
HellingerDist(x, y)
}
\arguments{
\item{x}{Required. a categorical variable.}

\item{y}{Required. a categorical variable.}
}
\value{
Hillinger Distance.
}
\description{
Calculate Hellinger Distance between two categorical variables
}
