% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAtlas.R
\name{makeAtlas}
\alias{makeAtlas}
\title{Make Atlas}
\usage{
makeAtlas(
  reference,
  npc = NULL,
  scale = FALSE,
  vis.atlas = T,
  col.by = NULL,
  colors = NULL,
  text = NULL,
  rank = TRUE,
  assay = "data"
)
}
\arguments{
\item{reference}{Required. Reference sce.}

\item{npc}{Default: 100. Number of Principal Components to build for the atlas.}

\item{scale}{Default: FALSE. Whether to scale the reference features.}

\item{vis.atlas}{Default: TRUE. Whether to visualize the atlas.}

\item{col.by}{Optional. Color reference samples by which metadata attribute.}

\item{colors}{Optional. Color Scheme of col.by. Should be a named vector using color codes as values and labels of 'col.by' as names.}

\item{text}{Optional. Additional annotation of reference samples.}

\item{rank}{Default: TRUE. Whether to perform rank transformation on the reference data.}

\item{assay}{Default: 'data'. On which reference assay we perform rank transformation.}
}
\value{
Reference sce.
}
\description{
Perform PCA on the ranked reference data to build atlas.
}
