% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{pca}
\alias{pca}
\title{PCA}
\usage{
pca(x, retx = TRUE, center = TRUE, scale. = FALSE, rank = NULL)
}
\arguments{
\item{x}{Required. data.}

\item{retx}{Default: TRUE.}

\item{center}{Default: TRUE.}

\item{scale.}{Default: FALSE.}

\item{rank}{Default: Find by the elbow method.}
}
\value{
pca object.
}
\description{
Standard principle component analysis
}
