% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postScale.R
\name{postScale}
\alias{postScale}
\title{Post imputation data scaling.}
\usage{
postScale(
  query,
  preImpAssay = "data",
  postImpAssay = "SincastImpData",
  plotTrend = TRUE,
  dologScale = TRUE,
  k = NULL,
  returnObsVarEst = FALSE
)
}
\arguments{
\item{query}{Required. Query sce.}

\item{preImpAssay}{Default: 'data'. Which assay in the query sce corresponds to the original data matrix.}

\item{postImpAssay}{Default: 'SincastImpData'. Which assay in the query sce corresponds to the imputed data matrix.}

\item{plotTrend}{Default: TRUE. Whether to plot the genewise mean versus variance trend estimation.}

\item{dologScale}{Default: TRUE. Whether to perform log scaling on the original data matrix (Which should be done if the imputation also performed logscaling)..}

\item{k}{Optional. Basis dimension for GAM.}

\item{returnObsVarEst}{Default: FALSE. Whether to return observational variance estimation.}
}
\value{
Query sce object.
}
\description{
Perform Sincast post imputation data scaling.
}
