% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{PCA projection}
\usage{
project(reference, query, assay = "SincastScaledData")
}
\arguments{
\item{reference}{Required. Reference sce with PCA performed.}

\item{query}{Required. Query sce.}

\item{assay}{Default: 'SincastScaledData'. The query assay to be projected onto the reference atlas.}
}
\value{
PC scores
}
\description{
Perform PCA projection on sce object.
}
