% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{projectDefault}
\alias{projectDefault}
\title{Default PCA projection}
\usage{
projectDefault(x, load, center = FALSE, scale = FALSE)
}
\arguments{
\item{x}{Required. A feature by observation data matrix.}

\item{load}{Required. Loading matrix.}

\item{center}{Default: FALSE Centering factors of the reference.}

\item{scale}{Default: FALSE. Scaling factors of the reference.}
}
\value{
PC scores
}
\description{
Perform PCA projection.
}
