% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomWalk.R
\name{randomWalk}
\alias{randomWalk}
\title{Diffusion random walk}
\usage{
randomWalk(sce, data, logScale = T, t = 3)
}
\arguments{
\item{sce}{Required. A sce object, on which sincastImp has been applied. The
diffusion operator compted by sincastImp will be applied to facilitate diffusion random
walk.}

\item{data}{Required. A gene by cell expression matrix to be diffused. Expect matched columns
between the sce object and the data.}

\item{logScale}{Default: TRUE. Whether to logscale the data.}

\item{t}{Default: 3. Diffusion time. The power of Markov transition matrix.}
}
\value{
A imputed data matrix
}
\description{
Applying the diffusion operator computed by sincastImp to impute data. Can be used
for imputing full data with diffusion opperators computed on filtered features.
}
