% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankTrans.R
\name{rankTrans}
\alias{rankTrans}
\title{Rank transformation}
\usage{
rankTrans(sce, assay = "data")
}
\arguments{
\item{sce}{Required. A sce object.}

\item{assay}{Default: 'data'. On which data assay we perform rank transformation.}
}
\value{
SingleCellExperiment Object.
}
\description{
This function rank transforms a specified data slot in sce, generates a separate
rank expression matrix stored in the rank slot of the sce.
}
