% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcTransform.R
\name{rcTransform}
\alias{rcTransform}
\title{Relative count normalization}
\usage{
rcTransform(sce, scale.factor = 1e+06)
}
\arguments{
\item{sce}{Required. A sce object containing count matrix.}

\item{scale.factor}{Default: 1e6. Will produce Count Per Million transformed data if set default.}
}
\value{
SingleCellExperiment Object.
}
\description{
This function normalizes count matrix in the sce object, generates a separate
relative count matrix stored in the data slot of the sce.
}
