% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SincastCapybara.R
\name{SincastCapybara}
\alias{SincastCapybara}
\title{Sincast Capybara cell score}
\usage{
SincastCapybara(
  reference,
  query,
  clusterid,
  k = 5,
  Npc = NULL,
  w = NULL,
  colDataPrefix = "Cb_"
)
}
\arguments{
\item{reference}{Required. Reference sce.}

\item{query}{Required. Query sce.}

\item{clusterid}{Required. Predict query cell identity based on the reference metadata attribute specified by 'clusterid'.}

\item{k}{Default: 5. Partition each reference cluster into k sub-clusters.}

\item{Npc}{Default: all. How many Principal components to use for searching the nearest reference cluster centroid to each query cell.}

\item{w}{Optional. Assign weights to genes in restricted linear regression. Can be either a numerical vector of the same length as the number of genes or be the names of a reference metadata attribute.}

\item{colDataPrefix}{Default: 'Cb_'. Prefix added to the column names of Capybara scores, which will be stored in the ColData of the query sce.}
}
\value{
Query sce object.
}
\description{
Using restricted linear regression to predict cell identity continum.
}
