% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sincastImp.R
\name{sincastImp}
\alias{sincastImp}
\title{Sincast imputation}
\usage{
sincastImp(
  query,
  dologScale = T,
  assay = "data",
  npc = 50,
  scale = F,
  k = 30,
  umapdist = TRUE,
  a = NULL,
  knn = NULL,
  doLaplacian = TRUE,
  norm = "probabilistic",
  t = 3,
  col.by = NULL,
  colors = NULL,
  text = NULL,
  vis.dc = TRUE,
  saveGrpah = FALSE
)
}
\arguments{
\item{query}{Required. Query sce.}

\item{dologScale}{Default: TRUE. Whether to log scale the query data.}

\item{assay}{Default: 'data'. On which assay we perform imputation.}

\item{npc}{Default: 50. How many Principal components to compute on the query data.}

\item{scale}{Default: FALSE. Whether to Scale the query data for PCA.}

\item{k}{Default: 30. k neighborhood to infer adaptive Gaussian kernel.}

\item{umapdist}{Default: TRUE. Whether to scale Euclidean distances to distances beyond nearest neighbors as in the UMAP algorithm.}

\item{a}{Default: log2(k) and log(k/(k-1)) when setting umapdist to TRUE and FALSE respectively.
a < 1 represents the probability of a cell communicating with its kth nearest neighbor.
a > 1 represents the sum of probabilities of a cell communicating with its k nearest neighbors.}

\item{knn}{Default: k. A cell can only be connected to knn neighbors when t is set 1.}

\item{doLaplacian}{Default: TRUE. Whether to perform Laplacian normalization on the affinity matrix.}

\item{norm}{Default: 'probabilistic'. How to symmetrize the affinity matrix. Default is Probabilistic t-norm. The other option is 'fuji' (Fuzzy Jaccard index).}

\item{t}{Default: 3. Diffusion time. The power of Markov transition matrix.}

\item{col.by}{Optional. Color query cells by which metadata attribute.}

\item{colors}{Optional. Color Scheme of col.by. Should be a named vector using color codes as values and labels of 'col.by' as names.}

\item{text}{Optional. Additional annotation of query cells.}

\item{vis.dc}{Default: TRUE. Whether to visualize the diffusion component extracted from the Markov transition matrix.}

\item{saveGrpah}{Default: FALSE. Whether to save the igraph model}
}
\value{
Query sce object.
}
\description{
Perform Sincast imputation.
}
