% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visProjection.R
\name{visProjection}
\alias{visProjection}
\title{Visualize projection}
\usage{
visProjection(
  reference,
  query,
  colReference.by = NULL,
  colQuery.by = NULL,
  referenceColors = NULL,
  queryColors = NULL,
  referenceText = NULL,
  queryText = NULL
)
}
\arguments{
\item{reference}{Required. Reference sce.}

\item{query}{Required. Query sce.}

\item{colReference.by}{Optional. Color reference samples by which reference metadata attribute.}

\item{colQuery.by}{Optional. Color quey samples by which query metadata attribute.}

\item{referenceColors}{Optional. Color Scheme of colReference.by Should be a named vector using color codes as values and labels of 'colReference.by' as names.}

\item{queryColors}{Optional. Color Scheme of colQuery.by Should be a named vector using color codes as values and labels of 'colQuery.by' as names.}

\item{referenceText}{Optional. Additional annotation of reference samples}

\item{queryText}{Optional. Additional annotation of query cells.}
}
\value{
A plotly object.
}
\description{
Using plotly 3D to visualize query projection on the reference atlas.
}
